refactor(Application) : LottoService, NumberGenerator 패키징 변경
