+++ src/main/java/lotto/domain/model/PrizeTier.java
@@ -4,21 +4,6 @@ package lotto.domain.model;
 import java.util.Arrays;
 
 public enum PrizeTier {
-//    SIX_MATCHES((matchCount, hasBonus) -> matchCount == 6, 2_000_000_000L),
-//    FIVE_WITH_BONUS_MATCHES((matchCount, hasBonus) -> matchCount == 5 && hasBonus, 30_000_000L),
-//    FIVE_MATCHES((matchCount, hasBonus) -> matchCount == 5 && !hasBonus, 1_500_000L),
-//    FOUR_MATCHES((matchCount, hasBonus) -> matchCount == 4, 50_000L),
-//    THREE_MATCHES((matchCount, hasBonus) -> matchCount == 3, 5_000L),
-//    NO_RANK((matchCount, hasBonus) -> false, 0L);
-//
-//    private final BiPredicate<Integer, Boolean> matchRule;
-//    private final long prizeMoney;
-//
-//    PrizeTier(BiPredicate<Integer, Boolean> matchRule, long prizeMoney) {
-//        this.matchRule = matchRule;
-//        this.prizeMoney = prizeMoney;
-//    }
-
     SIX_MATCHES(6, BonusType.ANY, 2_000_000_000L),
     FIVE_WITH_BONUS_MATCHES(5, BonusType.MATCHED, 30_000_000L),
     FIVE_MATCHES(5, BonusType.UNMATCHED, 1_500_000L),
